expose simplepattern and simplepatternsplit tokenizers
Fixes for code review

Take out admonition blocks in reference detail pages on
these tokenizers because Lucene's regexes are better protected
against being too complex or causing deep stacks.

Move these tokenizers to the common-analysis module because
that's where we're relocating code that depends on
lucene-analyzers-common

For #23363
