+++ docs/reference/analysis/tokenizers/simplepattern-tokenizer.asciidoc
@@ -19,17 +19,6 @@ For an explanation of the supported features and syntax, see <<regexp-syntax,Reg
 The default pattern is the empty string, which produces no terms. This
 tokenizer should always be configured with a non-default pattern.
 
-[WARNING]
-.Beware of Pathological Regular Expressions
-========================================
-
-A badly written regular expression could run very slowly or even throw a
-StackOverflowError and cause the node it is running on to exit suddenly.
-
-Read more about http://www.regular-expressions.info/catastrophic.html[pathological regular expressions and how to avoid them].
-
-========================================
-
 [float]
 === Configuration
