+++ docs/reference/analysis/tokenizers/simplepatternsplit-tokenizer.asciidoc
@@ -20,17 +20,6 @@ The default pattern is the empty string, which produces one term containing the
 full input. This tokenizer should always be configured with a non-default
 pattern.
 
-[WARNING]
-.Beware of Pathological Regular Expressions
-========================================
-
-A badly written regular expression could run very slowly or even throw a
-StackOverflowError and cause the node it is running on to exit suddenly.
-
-Read more about http://www.regular-expressions.info/catastrophic.html[pathological regular expressions and how to avoid them].
-
-========================================
-
 [float]
 === Configuration
