+++ modules/analysis-common/src/main/java/org/elasticsearch/analysis/common/SimplePatternSplitTokenizerFactory.java
@@ -17,30 +17,28 @@
  * under the License.
  */
 
-package org.elasticsearch.index.analysis;
+package org.elasticsearch.analysis.common;
 
 import org.apache.lucene.analysis.Tokenizer;
-import org.apache.lucene.analysis.pattern.SimplePatternTokenizer;
+import org.apache.lucene.analysis.pattern.SimplePatternSplitTokenizer;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.env.Environment;
 import org.elasticsearch.index.IndexSettings;
+import org.elasticsearch.index.analysis.AbstractTokenizerFactory;
 
-public class SimplePatternTokenizerFactory extends AbstractTokenizerFactory {
+public class SimplePatternSplitTokenizerFactory extends AbstractTokenizerFactory {
 
     private final String pattern;
 
-    public SimplePatternTokenizerFactory(IndexSettings indexSettings, Environment environment, String name, Settings settings) {
+    public SimplePatternSplitTokenizerFactory(IndexSettings indexSettings, Environment environment, String name, Settings settings) {
         super(indexSettings, name, settings);
 
         String pattern = settings.get("pattern", "");
-        if (pattern == null) {
-            throw new IllegalArgumentException("pattern is missing for [" + name + "] tokenizer of type 'simplepattern'");
-        }
         this.pattern = pattern;
     }
 
     @Override
     public Tokenizer create() {
-        return new SimplePatternTokenizer(pattern);
+        return new SimplePatternSplitTokenizer(pattern);
     }
 }
